[Embedding][Serve] Align qwen3 rope flag dtype with apache/tvm #19431
apache/tvm#19431 updates the masked prefill kernel's apply_rope
signature to T.int64. Match that on the mlc-llm side so qwen3
decoder embedding compiles against the upstream kernel once #19431
lands.

+++ python/mlc_llm/model/qwen3/qwen3_model.py
@@ -193,7 +193,7 @@ class Qwen3Attention(nn.Module):  # pylint: disable=too-many-instance-attributes
             args=[hidden_states],
         )
 
-        apply_rope_flag = tirx.IntImm("int32", 1)
+        apply_rope_flag = tirx.IntImm("int64", 1)
         rotary_emb = position_embedding.llama_rope_with_position_map(
             theta=self.rope_theta,
             scale=1.0,
